+++ src/nadir.h
@@ -25,7 +25,7 @@ typedef struct NonReentrantLock*  HNonReentrantLock;
 void                  Sleep(uint64_t timeout_us);
 typedef long volatile TAtomic32;
 long                  AtomicAdd32(TAtomic32* value, long amount);
-bool                  AtomicCAS32(TAtomic32* store, int32_t compare, int32_t value);
+bool                  AtomicCAS32(TAtomic32* store, long compare, long value);
 
 size_t            GetNonReentrantLockSize();
 HNonReentrantLock CreateLock(void* mem);
